Blood_supply/bloodbank_service/requirements.txt
Django==5.0.7
djangorestframework==3.15.1
kafka-python==2.0.2
django-environ==0.11.2
python-dateutil==2.8.2